refactor controller changes and make it clearer
